Ensures that every line ends with `\n` and updates readme
Previously, the output content was text which did not end with a `\n`
character. This is not correct according to the Posix standard. In
particular, if one were to count the lines using `wc`, then one would
obtain a value that if off by one. (Note here that `wc --lines` actually
counts the number of newline characters in a file.)

This update ensures that every line of output ends with a `\n` which
actually simplifies the program slightly.

A summary of the output of the program, in table form, has been added to
the readme. This may be easier to read then the output of the program.